@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:dicoding_pemula_flutter/views/white_noise_screen.dart';
 
 void main() {
   runApp(MyApp());
@@ -23,12 +24,13 @@ class MyApp extends StatelessWidget {
         primarySwatch: Colors.blue,
       ),
       home: MyHomePage(title: 'Flutter Demo Home Page'),
+      // home: MusicPlayerDetails(),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, required this.title}) : super(key: key);
 
   // This widget is the home page of your application. It is stateful, meaning
   // that it has a State object (defined below) that contains fields that affect
@@ -100,6 +102,7 @@ class _MyHomePageState extends State<MyHomePage> {
               '$_counter',
               style: Theme.of(context).textTheme.headline4,
             ),
+            MusicPlayerDetails(),
           ],
         ),
       ),
